requirements.txt
####### Requirements.txt #######

###### Requirements without version specifiers ######
pyyaml
pyautogui
datetime
typing

###### Requirements with version specifiers ######
numpy >= 1.18.5
matplotlib >= 3.3.4

###### Exmaples ######
# See https://www.python.org/dev/peps/pep-0440/#version-specifiers
# docopt == 0.6.1             # Version Matching. Must be version 0.6.1
# keyring >= 4.1.1            # Minimum version 4.1.1
# coverage != 3.5             # Version Exclusion. Anything except version 3.5
# Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*